Provider Porkbun: Fix wildcard domain modifications
Description
Porkbun does not properly handle wildcard domains with an HTML encoded *. By default the URL is encoded, and * becomes %2A which then causes Porkbun to not return the correct DNS entry for *.domain.com resulting in duplicate *.domain.com DNS entries and other unwanted behavior. This updates the logic to append the domain owner (subdomain) to the url after HTML encoding it, preventing the '*' or any other subdomain characters from being HTML encoded.

Test-Plan
Setup several domains with *.domain.tld and domain.tld dns A entries.
Set the values for some dns A entries to 127.0.0.1
Verified that DDNS-Updater was able to correctly update all entries with the invalid IP address set.